requirements.txt
openai
cx_Oracle
lxml
python-dotenv